discord-auction-bot/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp==3.9.1",
    "aiosqlite>=0.21.0",
    "discord-py>=2.5.2",
]
